+++ compose-lint-checks/src/main/java/slack/lint/compose/ComposeLintsIssueRegistry.kt
@@ -4,6 +4,7 @@ package slack.lint.compose
 
 import com.android.tools.lint.client.api.IssueRegistry
 import com.android.tools.lint.client.api.Vendor
+import com.android.tools.lint.detector.api.CURRENT_API
 import com.android.tools.lint.detector.api.Issue
 import com.google.auto.service.AutoService
 
@@ -17,8 +18,8 @@ class ComposeLintsIssueRegistry : IssueRegistry() {
       feedbackUrl = "https://github.com/slackhq/compose-lints/issues",
     )
 
-  override val api: Int = 15
-  override val minApi: Int = 15 // 8.2.0-alpha06
+  override val api: Int = CURRENT_API
+  override val minApi: Int = CURRENT_API
 
   @Suppress("SpreadOperator")
   override val issues: List<Issue> =